refactor: startClusterPreview thunk
